pyproject.toml
[project]
name = "splitposter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atproto>=0.0.54",
    "mastodon-py>=1.8.1",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
